% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{EBmlr3-package}
\alias{EBmlr3}
\alias{EBmlr3-package}
\title{EBmlr3: Wrapping EBM Learners as mlr3 Learners}
\description{
Wrapping EBM Learners as mlr3 Learners.
}
\author{
\strong{Maintainer}: Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})

}
